agents/requirements.txt
uagents
pydantic
requests
supabase
python-dotenv
python-dateutil

